Move UICollectionView extension to the new file
